+++ packages/plugins/i18n/meta.ts
