+++ packages/plugins/i18n/src/composable/useTranslate.ts
@@ -102,6 +102,7 @@ const ensureI18n = (obj, send) => {
 
     useCanvas().canvasApi.value?.setLocales(messages, true)
   } catch (e) {
+    throw new Error(e)
     // 不需要处理，有报错的词条会在画布初始化的时候统一调setLocales这个方法
   }
